restructured component folder
