+++ client/src/components/Buzzwords/Buzzwords.js
