+++ client/src/components/Buzzwords/Buzzwords.md
@@ -2,5 +2,5 @@
 This is a documentation about the buzzwords/skills.
 
 ```js
-<Buzzwords buzzwords={['One', 'Two', 'Three']}/>
+<Buzzwords buzzwords={["One", "Two", "Three"]} />
 ```
