+++ client/src/components/Header/Header.js
@@ -1,9 +1,9 @@
 import { Link } from "react-router-dom";
 import styled from "styled-components";
 import PropTypes from "prop-types";
-import { ReactComponent as BurgerIcon } from "../icons/burger.svg";
-import { ReactComponent as SearchIcon } from "../icons/search.svg";
-import { ReactComponent as HeartIcon } from "../icons/heart.svg";
+import { ReactComponent as BurgerIcon } from "../../icons/burger.svg";
+import { ReactComponent as SearchIcon } from "../../icons/search.svg";
+import { ReactComponent as HeartIcon } from "../../icons/heart.svg";
 
 export default function Header({ open, setOpen, headline }) {
   const openMenu = () => {
