+++ client/src/components/Searchbar/Searchbar.spec.js
