+++ client/src/pages/Profile.js
@@ -1,14 +1,10 @@
-import PropTypes from 'prop-types';
-import ProfileCard from '../components/ProfileCard';
+import PropTypes from "prop-types";
+import ProfileCard from "../components/ProfileCard/ProfileCard";
 
-
-export default function Profile({open}) {
-
-    return(
-        <ProfileCard open={open}/>
-    )
+export default function Profile({ open }) {
+  return <ProfileCard open={open} />;
 }
 
 Profile.propTypes = {
-    open: PropTypes.bool
-}
+  open: PropTypes.bool,
+};
